pyproject.toml
[project]
name = "harness-kit"
version = "0.1.0"
description = "A framework for Harness Engineering — constraints, feedback loops, and lifecycle systems that make AI coding agents reliable at scale."
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.9",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pathspec>=0.12.0",
    "platformdirs>=4.0",
]

[project.scripts]
harness = "harness_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/harness_cli"]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--strict-markers", "--tb=short"]
